requirements.txt
aws-cfn-bootstrap==1.4
awscli==1.16.102
Babel==0.9.4
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.8.0
boto==2.48.0
botocore==1.12.92
chardet==2.0.1
cloud-init==0.7.6
colorama==0.2.5
configobj==4.7.2
docutils==0.11
ecdsa==0.11
futures==3.0.3
hibagent==1.0.0
iniparse==0.3.1
Jinja2==2.7.2
jmespath==0.9.2
jsonpatch==1.2
jsonpointer==1.0
kitchen==1.1.1
lockfile==0.8
lxml==4.4.1
MarkupSafe==0.11
mysqlclient==1.4.4
nltk==3.4.5
nose==1.3.4
numpy==1.14.5
pandas==0.24.2
paramiko==1.15.1
PIL==1.1.6
ply==3.4
pyasn1==0.1.7
pycrypto==2.6.1
pycurl==7.19.0
pygpgme==0.3
pyliblzma==0.5.3
pystache==0.5.3
python-daemon==1.5.2
python-dateutil==2.8.1
python27-sagemaker-pyspark==1.2.4
pytz==2019.3
pyxattr==0.5.0
PyYAML==3.11
requests==1.2.3
rsa==3.4.1
simplejson==3.6.5
six==1.8.0
soupsieve==1.9.3
urlgrabber==3.10
urllib3==1.24.3
virtualenv==15.1.0
windmill==1.6
yum-metadata-parser==1.1.4
